Fix scenario when sending email to address that doesn't exist
It's a valid scenario to send an email to address that is not
within the system. We use DataRow to populate something similar
to user object and send it to the notification.

Most of the internal notifications are sent this way.

If we want to check if user is active before sending an email,
we should do that only if we find the user.

remp/helpdesk#503