Fix sourcing of RTM params to session from cookies
Originally the CRM used RTM params found in the URLs query parameters.
This was not the intention and sufficient, because the tracking attribution
might come from the cookie-based rtm_params which are set by remplib.js.

This change attempts to read RTM params from URL query params and if it's
not present, from cookie.

We also remove deprecated utm_params transition.

remp/crm#2354