Fix propagation of mail user preferences if they didn't exist yet
Mailer-CRM could got into state, where Mailer didn't have
the default user preferences created yet, but CRM wanted to update
them. CRM didn't correctly build the request, because it was based
on the existing user preferences provided by Mailer (which were empty).

The change request now operates with the submitted data and uses
preferences provided by Mailer as an optional feature. If they don't
exist, CRM assumes that the user IS NOT subscribed to the newsletter
and builds the change requests based on that information.

remp/crm#2883